Made a MenuBuilder class
To separate the menus.
