+++ Festivity/Utils/MenuBuilder.cs
@@ -1,0 +1,125 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Threading;
+
+namespace Festivity
+{
+    class MenuBuilder
+    {
+        public static List<MenuOption> MainMenu()
+        {
+            // Create List of menu options
+            List<MenuOption> newMenuOptions = new List<MenuOption>
+            {
+                new MenuOption("Register", () =>
+                {
+                    Console.Clear();
+                    UserRegisterPage.CreateUser();
+                }),
+                new MenuOption("Login", () =>
+                {
+                    Console.Clear();
+                    UserLoginPage.LoginPage();
+                }),
+                new MenuOption("Festivals", () =>
+                {
+                    Console.Clear();
+                    CatalogPage.CatalogMain();
+                }),
+                new MenuOption("Exit", () =>
+                {
+                    Environment.Exit(0);
+                })
+            };
+
+            return newMenuOptions;
+        }
+
+        public static List<MenuOption> MainMenuUser()
+        {
+            // Create List of menu options
+            List<MenuOption> newMenuOptions = new List<MenuOption>
+            {
+                new MenuOption("Festivals", () =>
+                {
+                    Console.Clear();
+                    CatalogPage.CatalogMain();
+                }),
+                new MenuOption("My Account", () =>
+                {
+                    Console.Clear();
+                    UserAccountPage.AccountPage();
+                }),
+                new MenuOption("My Tickets", () =>
+                {
+                    Console.Clear();
+                    TicketTableManager.Initiate();
+                }),
+                new MenuOption("Logout", () =>
+                {
+                    Console.Clear();
+                    Console.WriteLine("Successfully logged out!");
+                    Thread.Sleep(1000);
+                    Console.Clear();
+                    Menu.option = 0;
+                    LoggedInAccount.LogOut();
+                    Program.Main(); //new string[] { }
+                }),
+                new MenuOption("Exit", () =>
+                {
+                    Environment.Exit(0);
+                })
+            };
+
+            return newMenuOptions;
+        }
+
+        public static List<MenuOption> MainMenuOrganiser()
+        {
+            // Create List of menu options
+            List<MenuOption> newMenuOptions = new List<MenuOption>
+            {
+                new MenuOption("Festivals", () =>
+                {
+                    Console.Clear();
+                    CatalogPage.CatalogMain();
+                }),
+                new MenuOption("Register Festivals", () =>
+                {
+                    Console.Clear();
+                    Menu.option = 0;
+                    FestivalRegister.activeScreen = true;
+                    FestivalRegister.currentRegisterSelection = "Main";
+                    FestivalRegister.ShowFestivalRegister();
+                    Thread.Sleep(1000);
+                }),
+                new MenuOption("My Account", () =>
+                {
+                    Console.Clear();
+                    UserAccountPage.AccountPage();
+                }),
+                new MenuOption("My Tickets", () =>
+                {
+                    Console.Clear();
+                    TicketTableManager.Initiate();
+                }),
+                new MenuOption("Logout", () =>
+                {
+                    Console.Clear();
+                    Console.WriteLine("Successfully logged out!");
+                    Thread.Sleep(1000);
+                    Console.Clear();
+                    Menu.option = 0;
+                    LoggedInAccount.LogOut();
+                    Program.Main(); //new string[] { }
+                }),
+                new MenuOption("Exit", () =>
+                {
+                    Environment.Exit(0);
+                })
+            };
+
+            return newMenuOptions;
+        }
+    }
+}
